Read and write simple-font text through its own encoding
A user editing a form's date field saw the content tool report the text
as `-=>-/>?-?@` where the page reads `05/08/2026`, and edits that
"changed a lot of content on the page". Two independent bugs.

**A simple font's byte is not its character.** `PdfPageElements.of`
decoded a non-/Type0 show string with `latin1.decode`, and
`_SimpleRunCodec` mirrored that: matching on `String.fromCharCode(code)`
and writing replacements as `replace.codeUnits`. That only holds when the
font's codes happen to be Latin-1. Section 9.6.6 lets a document remap
every code through `/Encoding` `/Differences`, and a subsetted font
routinely renumbers its codes from scratch - the reporter's font drew
`05/08/2026` from the bytes `2D 3D 3E 2D 2F 3E 3F 2D 3F 40`, which is
exactly `-=>-/>?-?@` in Latin-1. Their "B = 8" note is the same table read
backwards: typing `B` wrote `0x42`, and `0x42` is whatever glyph the
subset put there.

The renderer never had this problem (`PdfFontInfo.charFor` resolves
/ToUnicode, /Differences, the built-in Symbol/ZapfDingbats encodings, then
the base encoding), which is why selection and search read correctly the
whole time. Only the content editor had its own shortcut.

Adds `SimpleFont`, the simple-font sibling of `Type0Font`: one place that
resolves a code to text the way the renderer does, plus the reverse table
for re-encoding replacements and the `/Widths` lookup (which absorbs
`content_editor`'s `_widthsFor`). Used by `PdfPageElements`, both run
codecs, the `'`/`"` operator path, and reflow's `_SimpleReflowFont`.

Encodability is deliberately narrower than decodability: decoding keeps
the lenient Latin-1 fallback, but a replacement may only use codes the
font *declares* (a glyph name from the base encoding or /Differences, or a
/ToUnicode entry), because a subset that dropped `9` has no glyph at
`0x39`. An undrawable replacement leaves the run untouched - the contract
`Type0Font` already had. A font that declares no encoding at all (base-14,
and everything this library writes) keeps the identity in both
directions, so existing round-trips are unchanged.

Fixing the decode also created a latent bug in paragraph reflow, fixed
here: `content_reflow` reads `element.text`, so once that decoded
correctly a remapped font would start matching, and `showOp` would have
written `text.codeUnits` as bytes.

**"It changes it everywhere."** Separately, the element strip's "Replace
text" went through the page-wide `PdfEditor.replaceText`. The targeted
`replaceElementText` already existed (it backs the selection-menu edit);
this path just never used it. Both element-strip entry points now share
one helper that uses the targeted API.

`SimpleFont` needs the Adobe glyph-name tables, which lived one layer up
in `pdf_graphics`. They are pure data, so they move down to
`pdf_document/src/fonts/encodings.dart`; the old path is now a `show`-list
re-export so the font engine's imports are untouched.
