+++ packages/pdf_document/lib/pdf_document.dart
@@ -15,6 +15,7 @@ export 'src/disk_cache.dart';
 export 'src/document.dart';
 export 'src/editor.dart';
 export 'src/font_embedder.dart';
+export 'src/fonts/encodings.dart';
 export 'src/form.dart';
 export 'src/fulcio.dart';
 export 'src/image.dart';
