+++ packages/pdf_document/lib/src/editor.dart
@@ -23,6 +23,7 @@ import 'page.dart';
 import 'page_labels.dart';
 import 'rect.dart';
 import 'signing_identity.dart';
+import 'simple_font.dart';
 import 'stamp_template.dart';
 import 'struct_tree.dart';
 import 'takeoff.dart';
